specs/demo/requirements.txt
# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 环境变量与配置
python-dotenv>=1.0.0

# 测试与AI/数据
pytest>=7.0.0
openai>=1.0.0
tinyshare>=0.1.0
requests>=2.31.0

# 可视化
plotly>=5.0.0

# Cloudflare R2（S3 兼容）
boto3>=1.34.0

